Fixed only showing one item bug and search not clearing previous results bug
Instead of changing the innerHTML of the phoneDiv, the innerHTML of phonesContainer was being changed previously, which made sure that only one item was being displayed, therefore fixed it.

Secondly, the phonesContainer's old items were not removed before displaying new items. It was commented out for some reason, therefore uncommenting it fixed it